+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Tested on build 22000.348.
 * Added some more debug messages
 * Fixed some rendering problems when themes are disabled
 * Fixed regression of [#161](https://github.com/valinet/ExplorerPatcher/issues/161#issuecomment-986234002) (.1)
+* Possibility to disable per-application window lists (`Alt`+`) ([#283](https://github.com/valinet/ExplorerPatcher/issues/283#issuecomment-986261712)) (.2)
 
 ## 22000.348.39